requirements.txt
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.5
matplotlib>=3.4.3
scikit-learn>=0.24.2
pytorch-grad-cam>=1.3.7
pandas>=1.3.0
Pillow>=8.2.0
tqdm>=4.60.0
torchmetrics>=0.5.0
